add representation and revert some changes
