+++ express/common.py
@@ -2,7 +2,6 @@
 and pipelines."""
 
 import enum
-import types
 import typing as t
 
 
@@ -55,44 +54,3 @@ class Field(t.NamedTuple):
 
     name: str
     type: Type
-
-
-class Subset:
-    """
-    Class representing an Express subset.
-
-    Args:
-        specification: The part of the manifest json representing the subset
-        base_path: The base path which the subset location is defined relative to
-    """
-
-    def __init__(self, specification: dict, *, base_path: str) -> None:
-        self._specification = specification
-        self._base_path = base_path
-
-    @property
-    def location(self) -> str:
-        """The resolved location of the subset"""
-        return self._base_path.rstrip("/") + self._specification["location"]
-
-    @property
-    def fields(self) -> t.Mapping[str, Field]:
-        """The fields of the subset returned as a immutable mapping."""
-        return types.MappingProxyType(
-            {
-                name: Field(name=name, type=field["type"])
-                for name, field in self._specification["fields"].items()
-            }
-        )
-
-    def add_field(self, name: str, type_: Type) -> None:
-        if name in self._specification["fields"]:
-            raise ValueError("A field with name {name} already exists")
-
-        self._specification["fields"][name] = {"type": type_.value}
-
-    def remove_field(self, name: str) -> None:
-        del self._specification["fields"][name]
-
-    def __repr__(self) -> str:
-        return f"{self.__class__.__name__}({self._specification!r}"
